添加固件传输进度条窗口；添加点击按钮固件解压模块；解决固件传送出现同一条数据发送多次得问题；添加文件数据传送得重发机制；解决连续发送多个固件出现的问题。

+++ checkout.h
@@ -40,7 +40,7 @@ static const unsigned short crc16tab[256]= {
 
 //计算CRC16
 quint16 crc16_ccitt(const char *buf, int len);
-uint8_t TX_CheckSum(uint8_t *buf, uint8_t len);
-uint8_t RX_CheckSum(uint8_t *buf, uint8_t len);
+uint8_t TX_CheckSum(uint8_t *buf, uint16_t len);
+uint8_t RX_CheckSum(uint8_t *buf, uint16_t len);
 
 #endif // CHECKOUT_H
